move out tensorboard progress bar
